Add Changes

+++ src/main/java/com/tw/apistackbase/service/CompanyService.java
@@ -3,50 +3,63 @@ package com.tw.apistackbase.service;
 import com.tw.apistackbase.core.Company;
 import com.tw.apistackbase.repository.CompanyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class CompanyService {
     @Autowired
     CompanyRepository companyRepository;
 
-    public List<Company> findAll()
-    {
+    public Iterable<Company> list(Integer page, Integer pageSize) {
+        return companyRepository.findAll(PageRequest.of(page-1,pageSize));
+    }
+
+    public List<Company> getAll() {
         return companyRepository.findAll();
     }
 
-    public Page<Company> findAll(Integer page, Integer pageSize)
-    {
-        return companyRepository.findAll(PageRequest.of(page - 1 ,pageSize, Sort.by("name").descending()));
-    }
-
-    public Company save(Company company) {
-        return companyRepository.save(company);
-    }
-
-    public List<Company> findByFirstnameLike(String name) {
+    public List<Company> getCompanyLikeName(String name) {
+        if(name == null) {
+            return companyRepository.findAll();
+        }
         return companyRepository.findByFirstnameLike(name);
     }
 
-    public Optional<Company> findById(Long id) {
-        return companyRepository.findById(id);
+    public Boolean deleteCompanyById(Long id){
+        if ( companyRepository.findById(id).isPresent()) {
+            companyRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
-    public void deleteById(Long id) {
-        companyRepository.deleteById(id);
+    public Company add(Company company) {
+        companyRepository.save(company);
+        return company;
     }
 
-    public void deleteAll() {
+    public Boolean deleteAllCompany(){
         companyRepository.deleteAll();
+        if (companyRepository.findAll().isEmpty()){
+            return true;
+        }
+        return false;
+
     }
 
-    public Company getOne(Long id) {
-        return companyRepository.getOne(id);
+    public Company modifyCompany(Long id, Company company){
+        if ( companyRepository.findById(id).isPresent()) {
+            Company company1 = companyRepository.getOne(id);
+            company1.setName(company.getName());
+            companyRepository.save(company1);
+            return company1;
+        } else {
+            return null;
+        }
     }
+
+
 }
